+++ dataherb/deprecation/cmd/configs.py
@@ -9,7 +9,10 @@ logger.add(sys.stderr, level="INFO", enqueue=True)
 
 
 def load_dataherb_config(config_path=None, no_config_error=True):
-    """[Deprecated]
+    """
+
+    [Deprecated] Use `dataherb.utils.configs.Config` instead.
+
     Loads the dataherb config file.
 
     Load the content from the specified file as the config. The config file has to be json.
